code updated till section 12, pending deployment
