+++ backend/app.js
@@ -8,7 +8,8 @@ const userRoutes = require("./routes/user");
 
 const app = express();
 
-mongoose.connect("mongodb+srv://meancourse:meancourse@cluster0.fscz4.mongodb.net/meancoursedb?retryWrites=true&w=majority")
+
+mongoose.connect(`mongodb+srv://meancourse:${process.env.MONGO_ATLAS_PW}@cluster0.fscz4.mongodb.net/meancoursedb?retryWrites=true&w=majority`)
     .then(() => {
         console.log(`Connected to Mongo DB`);
     })
