+++ backend/middleware/file-path-fetch.js
@@ -1,0 +1,25 @@
+const multer = require("multer"); //requires configuration
+
+const MIME_TYPE_MAP = {
+    "image/png": "png",
+    "image/jpeg": "jpeg",
+    "image/jpg": "jpg",
+};
+
+const storage = multer.diskStorage({
+    destination: (req, file, cb) => {
+        const isValid = MIME_TYPE_MAP[file.mimetype];
+        let error = new Error("Invalid mime type");
+        if (isValid) {
+            error = null;
+        }
+        cb(error, "backend/images"); //path is related to server.js path
+    },
+    filename: (req, file, cb) => {
+        const name = file.originalname.toLowerCase().split(" ").join("-");
+        const ext = MIME_TYPE_MAP[file.mimetype];
+        cb(null, name + "-" + Date.now() + "." + ext);
+    },
+});
+
+module.exports = multer({ storage: storage, }).single("image")
