+++ nodemon.json
@@ -1,0 +1,6 @@
+{
+    "env":{
+        "MONGO_ATLAS_PW": "meancourse",
+        "JWT_KEY":"this_is_the_secret_key_which_is_used_to_generate_token"
+    }
+}
