+++ src/app/angular-material.module.ts
@@ -1,0 +1,39 @@
+import { NgModule } from "@angular/core";
+
+import { MatInputModule } from '@angular/material/input';
+import { MatCardModule } from '@angular/material/card';
+import { MatButtonModule } from '@angular/material/button';
+import { MatToolbarModule } from '@angular/material/toolbar';
+import { MatExpansionModule } from '@angular/material/expansion';
+import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
+import { MatPaginatorModule } from "@angular/material/paginator";
+import { MatDialogModule } from "@angular/material/dialog";
+
+@NgModule({
+    // imports: [
+    //     MatInputModule,
+    //     MatCardModule,
+    //     MatButtonModule,
+    //     MatToolbarModule,
+    //     MatExpansionModule,
+    //     MatProgressSpinnerModule,
+    //     MatPaginatorModule,
+    //     MatDialogModule
+    // ], //by default they are not exposed to any other module
+    //to make usable we have add another key
+    exports: [
+        MatInputModule,
+        MatCardModule,
+        MatButtonModule,
+        MatToolbarModule,
+        MatExpansionModule,
+        MatProgressSpinnerModule,
+        MatPaginatorModule,
+        MatDialogModule
+    ]
+})//this turns class into angular module
+//this module will import things from angular material and export them again
+//we can remove the imports and just keep the exports, importing will be done automatically
+export class AngularMaterialModule {
+
+}
