+++ src/app/auth/auth.module.ts
@@ -1,0 +1,27 @@
+import { NgModule } from "@angular/core";
+import { RouterModule } from "@angular/router";
+import { FormsModule } from '@angular/forms';
+import { CommonModule } from "@angular/common";
+
+import { LoginComponent } from './login/login.component';
+import { SignupComponent } from './signup/signup.component';
+import { AngularMaterialModule } from "../angular-material.module";
+import { AuthRoutingModule } from "./auth-routing.module";
+
+
+@NgModule({
+    declarations: [
+        LoginComponent,
+        SignupComponent
+    ],
+    imports: [
+        CommonModule,
+        AngularMaterialModule,
+        RouterModule,
+        FormsModule,
+        AuthRoutingModule
+    ]
+})
+export class AuthModule {
+
+}
